requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==0.61.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
idna==3.10
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
pytesseract==0.3.13
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
